Move LoadNetworkSpinner to spinner.js for reuse
